Initial work on supporting failover
Add a user specified failover timeout that will allow tasks to keep
running even if the Scheduler dies. Note that failover will only occur
in cases where the machine running the framework goes down suddenly.
If stolos or mesos throw an exception the Scheduler will be shut down cleanly,
and no failover will occur.

The current implementation uses the Stolos queueback as persistent
storage to keep track of the framework_id of failed Scheduler. This
might lead to odd behaviour when more than one instance of the framework
is running in Marathon (which shouldn't happen anyway).